Complete phase 3
A trie can now be populated with a dictionary list of words

Bug fix:

The previous suggestions implementation contained a bug where the
suggestions given was an incomplete list, cut short if a suggestion was
a shorter word than the previous computed suggestion (e.g. if the user
typed in "ap" and our dictionary had both "apple" and "ape" as already
inserted words, only "apple" would be returned as a suggestion since
"ape" is shorter than "apple" and comes after "apple" in the suggestion
computation).

To problem was that I was returning the recursive `getSuggestions`
function if the current node was not a leaf node (i.e. end of an
inserted word) which would incorrectly prevent suggestions from being
added to the `suggestions` array. To fix this bug, I could simply remove
the return statement for the `getSuggestions` recursive call, allowing
the for loop to continue.