Increase performance of loop to find the end of query node
In order to see if a trie contains a node path for the given query, we
split the query and iterate through each letter to see if there exists a
child node with the value of each successive letter. The solution
implemmented before this commit used `Array#reduce` to loop over each
letter in the query. This works fine, but means we have no way to
short-circuit the loop should we find that there exists no node path
somewhere in the middle of the query's letters; the loop continues until
it reaches the end of the array, i.e. the last letter in the query.

This commit refactors the loop to search for the query node path to
include a short circuit should we encounter a node with no child.