Replace getEndOfQueryNode with getNodePath
getEndOfQueryNode was only helpful insofar as it allowed me to find the
starting point of discovering suggestions. However, it was limited given
that I couldn't retrace the node path back up the trie, something that
will be helpful when a deletion method is introduced.

The new method getNodePath returns an array of nodes for a given string
query, where each preceding node is the parent of node after it, all the
way back up to the root node.