Fix issue of bad suggestions when query matches start of existing word
The private function getNodePath had a subtle bug in that it would still
return a list of nodes even if that list did not have a one-to-one
mapping of nodes to letters in the query. This resulted in suggestions
being populated with words if the query had *at least* one letter prefix
that matched the word, whereas it should only match if the query has
*exactly* a prefix of a word (meaning that queries longer than existing
words should also return an empty suggestions list).

Fixes #1