Add function to generate hierarchical tree object
As the code stands, there are no public properties exported from the
createTrie factory function; meaning that getter methods must be used
to access data at each level in the tree. This makes it hard to export
the trie as a consumable object, such as for visualization libraries
like D3.

The new public method in this commit will create a hierarchy object that
is compliant with D3's requests.

refs:
- https://github.com/d3/d3-hierarchy#d3-hierarchy